ARC systems now using conda >4.4:
Prior to conda 4.4 the way to set up your env was to add the
conda_root/bin to the path, then change environments by sourcing the
activate script.

Starting with conda v4.4 conda comes with a script that should be
sourced in the bashrc located at conda_root/etc/conda.sh. Changing
environments are now done through the activate and deactivate conda
sub_commands.

TODO:
I will need to make the same changes on tantiveIV when conda is updated
there.